+++ framework/ui/components/organisms/vc-gallery/_internal/vc-gallery-preview/vc-gallery-preview.vue
@@ -17,7 +17,7 @@
       <div class="tw-box-border p-5 tw-grow tw-basis-0 tw-w-full">
         <div
           class="bg-contain tw-bg-no-repeat tw-bg-center tw-w-full tw-h-full tw-box-border"
-          :style="{ backgroundImage: 'url(' + currentImage.url + ')' }"
+          :style="imageHandler"
         ></div>
         <div
           v-if="localIndex > 0"
@@ -30,7 +30,7 @@
           ></VcIcon>
         </div>
         <div
-          v-if="localIndex < unref(images).length - 1"
+          v-if="localIndex < images.length - 1"
           class="tw-absolute tw-top-2/4 -tw-mt-9 tw-h-[72px] tw-w-[72px] tw-flex tw-items-center tw-justify-center tw-rounded-full tw-bg-[#f1f6fa] tw-cursor-pointer tw-text-[#a1c0d4] [--icon-size-xl: 36px] hover:tw-shadow-[0_0_20px_rgba(31,40,50,0.15)] tw-right-[25px]"
           @click="localIndex++"
         >
@@ -42,7 +42,7 @@
       </div>
       <div class="tw-p-4 tw-pb-[40px] tw-max-w-full tw-overflow-x-auto tw-box-border tw-shrink tw-flex">
         <div
-          v-for="(item, i) in unref(images)"
+          v-for="(item, i) in images"
           :key="i"
           class="tw-m-1 tw-opacity-60"
           :class="{
@@ -63,14 +63,14 @@
 </template>
 
 <script lang="ts" setup>
-import { computed, ref, ComputedRef, MaybeRef, unref } from "vue";
+import { computed, ref } from "vue";
 import { VcPopup, VcLink, VcIcon, VcImage } from "../../../../";
 import { IImage } from "./../../../../../../core/types";
 import { useI18n } from "vue-i18n";
 
 export interface Props {
-  images?: MaybeRef<IImage[]>;
-  index: number | ComputedRef<number>;
+  images?: IImage[];
+  index: number;
 }
 
 export interface Emits {
@@ -85,12 +85,15 @@ const props = withDefaults(defineProps<Props>(), {
 });
 const { t } = useI18n({ useScope: "global" });
 const localIndex = ref(props.index);
-const currentImage = computed(() => {
-  if (unref(props.images) && unref(props.images).length) {
-    return unref(props.images)[localIndex.value];
+
+const imageHandler = computed(() => {
+  if (currentImage.value.url) {
+    return `background-image: url(${CSS.escape(currentImage.value.url)})`;
   }
   return undefined;
 });
+
+const currentImage = computed(() => props.images[localIndex.value]);
 
 const copyLink = (link: string) => {
   if (link.charAt(0) === "/") {